Remove BusbarSectionFinderTraverser (now in powsybl-core) Fix unit tests (after properties added on limits in powsybl-core)
