+++ src/test/java/org/gridsuite/network/map/dto/mapper/BranchInfosMapperTest.java
@@ -9,10 +9,10 @@ import org.assertj.core.api.junit.jupiter.SoftAssertionsExtension;
 import org.gridsuite.network.map.dto.common.CurrentLimitsData;
 import org.gridsuite.network.map.dto.common.CurrentLimitsData.Applicability;
 import org.gridsuite.network.map.dto.common.TemporaryLimitData;
+import org.gridsuite.network.map.dto.utils.CurrentLimitsDtoTest;
 import org.gridsuite.network.map.dto.utils.MockDto;
-import org.gridsuite.network.map.dto.utils.MockDto.CurrentLimitsDtoTest;
-import org.gridsuite.network.map.dto.utils.MockDto.OperationalLimitsGroupDtoTest;
-import org.gridsuite.network.map.dto.utils.MockDto.TemporaryLimitDtoTest;
+import org.gridsuite.network.map.dto.utils.OperationalLimitsGroupDtoTest;
+import org.gridsuite.network.map.dto.utils.TemporaryLimitDtoTest;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
